fix migration and init db, update repository
