+++ FALOFinancialProofing/Models/FALOFinancialProofingDbContext.cs
@@ -9,7 +9,6 @@ namespace FALOFinancialProofing.Models
 
         public DbSet<User> Users { get; set; }
         public DbSet<Role> Roles { get; set; }
-        //public DbSet<UserRole> UserRoles { get; set; }
         public DbSet<RefreshToken> RefreshTokens { get; set; }
 
         #endregion DBSet
